November 24 2020 Update
Resolved a bug in the Update feature where if you attempted to preform play list download would fail because youtube-dl.exe is not up to date and then try to update it without clearing the url would fail. Provided a pop up message box to let you know to update to preform that function and then you can continue to download without having to input the url again.
Added a comment metadata tag which contains the url of the video so you know where it came from if you needed to find it again.